Now the start_staging_gfz script uses docker-compose
As I had trouble running the updated nuxt version in my
old node version, I adopted the existing docker-compose file
to be used with the addresses of the staging vms for gfz.

I also needed to write an explicit docker file with the installation
of python3 and the build-essential package.