overwrite the default.conf, so that the forwarding works
Before this commit, the nginx container still used its default.conf,
which listens on port 80 and tries to find all uris as static files.

With this change, the default.conf will be overwritten, and all
requests on port 80 will be redirected to port 443.

(But it doesn't support static files anymore).